+++ cpp-api-client/src/cpprest-client/model/LogTree.cpp
@@ -28,14 +28,14 @@ LogTree::LogTree()
     m_ChildrenIsSet = false;
     m_CreatedAt = 0L;
     m_CreatedAtIsSet = false;
-    m_Id = 0;
+    m_Id = 0L;
     m_IdIsSet = false;
     m_OriginIsSet = false;
-    m_ParentLogId = 0;
+    m_ParentLogId = 0L;
     m_ParentLogIdIsSet = false;
     m_Replies = 0L;
     m_RepliesIsSet = false;
-    m_RootLogId = 0;
+    m_RootLogId = 0L;
     m_RootLogIdIsSet = false;
     m_RunsIsSet = false;
     m_SubsystemsIsSet = false;
@@ -180,7 +180,7 @@ bool LogTree::fromJson(const web::json::value& val)
         const web::json::value& fieldValue = val.at(utility::conversions::to_string_t("id"));
         if(!fieldValue.is_null())
         {
-            int32_t refVal_id;
+            int64_t refVal_id;
             ok &= ModelBase::fromJson(fieldValue, refVal_id);
             setId(refVal_id);
         }
@@ -200,7 +200,7 @@ bool LogTree::fromJson(const web::json::value& val)
         const web::json::value& fieldValue = val.at(utility::conversions::to_string_t("parentLogId"));
         if(!fieldValue.is_null())
         {
-            int32_t refVal_parentLogId;
+            int64_t refVal_parentLogId;
             ok &= ModelBase::fromJson(fieldValue, refVal_parentLogId);
             setParentLogId(refVal_parentLogId);
         }
@@ -220,7 +220,7 @@ bool LogTree::fromJson(const web::json::value& val)
         const web::json::value& fieldValue = val.at(utility::conversions::to_string_t("rootLogId"));
         if(!fieldValue.is_null())
         {
-            int32_t refVal_rootLogId;
+            int64_t refVal_rootLogId;
             ok &= ModelBase::fromJson(fieldValue, refVal_rootLogId);
             setRootLogId(refVal_rootLogId);
         }
@@ -406,7 +406,7 @@ bool LogTree::fromMultiPart(std::shared_ptr<MultipartFormData> multipart, const 
     }
     if(multipart->hasContent(utility::conversions::to_string_t("id")))
     {
-        int32_t refVal_id;
+        int64_t refVal_id;
         ok &= ModelBase::fromHttpContent(multipart->getContent(utility::conversions::to_string_t("id")), refVal_id );
         setId(refVal_id);
     }
@@ -418,7 +418,7 @@ bool LogTree::fromMultiPart(std::shared_ptr<MultipartFormData> multipart, const 
     }
     if(multipart->hasContent(utility::conversions::to_string_t("parentLogId")))
     {
-        int32_t refVal_parentLogId;
+        int64_t refVal_parentLogId;
         ok &= ModelBase::fromHttpContent(multipart->getContent(utility::conversions::to_string_t("parentLogId")), refVal_parentLogId );
         setParentLogId(refVal_parentLogId);
     }
@@ -430,7 +430,7 @@ bool LogTree::fromMultiPart(std::shared_ptr<MultipartFormData> multipart, const 
     }
     if(multipart->hasContent(utility::conversions::to_string_t("rootLogId")))
     {
-        int32_t refVal_rootLogId;
+        int64_t refVal_rootLogId;
         ok &= ModelBase::fromHttpContent(multipart->getContent(utility::conversions::to_string_t("rootLogId")), refVal_rootLogId );
         setRootLogId(refVal_rootLogId);
     }
@@ -559,12 +559,12 @@ void LogTree::unsetCreatedAt()
 {
     m_CreatedAtIsSet = false;
 }
-int32_t LogTree::getId() const
+int64_t LogTree::getId() const
 {
     return m_Id;
 }
 
-void LogTree::setId(int32_t value)
+void LogTree::setId(int64_t value)
 {
     m_Id = value;
     m_IdIsSet = true;
@@ -599,12 +599,12 @@ void LogTree::unsetOrigin()
 {
     m_OriginIsSet = false;
 }
-int32_t LogTree::getParentLogId() const
+int64_t LogTree::getParentLogId() const
 {
     return m_ParentLogId;
 }
 
-void LogTree::setParentLogId(int32_t value)
+void LogTree::setParentLogId(int64_t value)
 {
     m_ParentLogId = value;
     m_ParentLogIdIsSet = true;
@@ -639,12 +639,12 @@ void LogTree::unsetReplies()
 {
     m_RepliesIsSet = false;
 }
-int32_t LogTree::getRootLogId() const
+int64_t LogTree::getRootLogId() const
 {
     return m_RootLogId;
 }
 
-void LogTree::setRootLogId(int32_t value)
+void LogTree::setRootLogId(int64_t value)
 {
     m_RootLogId = value;
     m_RootLogIdIsSet = true;
