+++ cpp-api-client/src/cpprest-client/model/Tag.h
@@ -65,11 +65,11 @@ public:
     /// <summary>
     /// The unique identifier of this entity.
     /// </summary>
-    int32_t getId() const;
+    int64_t getId() const;
     bool idIsSet() const;
     void unsetId();
 
-    void setId(int32_t value);
+    void setId(int64_t value);
 
     /// <summary>
     /// The label value of the tag.
@@ -93,7 +93,7 @@ public:
 protected:
     int64_t m_CreatedAt;
     bool m_CreatedAtIsSet;
-    int32_t m_Id;
+    int64_t m_Id;
     bool m_IdIsSet;
     utility::string_t m_Text;
     bool m_TextIsSet;
